Added window system and improved drag of phone-island (#28)
The nethlink page is now resizable, movable, minimisable and closable. when the nethlink page is closed, the user can reopen it with the context menu of the tray icon. The tray icon now offers a context menu with the possibility to exit the application (safeQuit) or to show/hide the current page (login or nethlink). 

* feat: add window frame
* fix phoneIsland positioning
* feat: storage of phoneisland position
* fix: now dragend doesn't click if the drag was started over a button
* fix: changed setPosition function to setBounds to solve the problem of changing window size while dragging
* fix: macos maximize now is closable
* fix: window positions memory
* add safeQuit on OS shutdown

Co-authored-by: Edoardo Spadoni <edoardo.spadoni@gmail.com>